feat: logging
